requirements.txt
mitmproxy
pyyaml
pytest-shutil